--- a/cli/flox-rust-sdk/src/providers/git.rs
+++ b/cli/flox-rust-sdk/src/providers/git.rs
@@ -250,9 +250,11 @@ impl GitCommandProvider {
         options: GitCommandOptions,
         path: P,
     ) -> Result<Self, GitCommandOpenError> {
+        debug!("attempting to open repo: path={}", path.as_ref().display());
         let bare = Self::is_bare_repo(&path)?;
 
         // resolved and canonicalized path to the git repo
+        debug!("determining path to git repo");
         let resolved_path = {
             let toplevel_or_git_dir = if bare {
                 let mut command = options.new_command();
@@ -280,6 +282,7 @@ impl GitCommandProvider {
                 .canonicalize()
                 .map_err(GitCommandOpenError::Canonicalize)?
         };
+        debug!("got non-canonical path: path={}", resolved_path.display());
 
         let path = path
             .as_ref()
@@ -289,6 +292,7 @@ impl GitCommandProvider {
         if resolved_path != path {
             return Err(GitCommandOpenError::Subdirectory);
         }
+        debug!("canonicalized path: path={}", path.display());
 
         Ok(GitCommandProvider {
             options,