--- a/pkgdb/include/flox/buildenv/command.hh
+++ b/pkgdb/include/flox/buildenv/command.hh
@@ -27,6 +27,19 @@ namespace flox::buildenv {
 
 /* -------------------------------------------------------------------------- */
 
+/**
+ * @class flox::buildenv::SystemNotSupportedByLockfile
+ * @brief An exception thrown when a lockfile is is missing a package.<system>
+ * entry fro the requested system.
+ * @{
+ */
+FLOX_DEFINE_EXCEPTION( BuildenvInvalidArguments,
+                       EC_BUILDENV_ARGUMENTS,
+                       "invalid arguments to buildenv" )
+/** @} */
+
+/* -------------------------------------------------------------------------- */
+
 /** @brief Evaluate and build a locked environment. */
 class BuildEnvCommand : NixState
 {
@@ -37,6 +50,7 @@ private:
   nlohmann::json             lockfileContent;
   std::optional<std::string> outLink;
   std::optional<System>      system;
+  std::optional<std::string> storePath;
   bool                       buildContainer;