## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be resolved. -->
This PR is a dramatic speedup to all operations that build environments (install, pull, etc). This is accomplished via two major changes. It's recommended to review this by each commit since the two major changes are each isolated to their commits.

### Reusing temporary builds
The first major change is reusing temporary builds of environments. During transactions a copy of the environment is built in a temporary directory so as to not modify the existing environment until we know the new one builds successfully. The result of the build is a Nix store path for the newly built environment. Previously this store path was discarded. After building the temporary copy, the environment is built again from scratch so that it can create the GC root in the correct location. Now this store path is retained and passed to `pkgdb buildenv` so that it can skip the build and simply set the GC root.

Original build timing:
- First build (temporary copy): 8.3s
- Second build: 4.9s

Passing the store path:
- First build (temporary copy): 8.8s
- Second build (only creating the GC root): 0.05s

### Using the eval cache

The second major change is using the eval cache during `pkgdb buildenv`. Using the eval cache is an opt-in mechanism, and this was simply not implemented previously. The first time a flake attribute is evaluated _anywhere_ the result is cached and stored in a SQLite database in `~/.cache/nix/eval-cache-v5`. Nix evaluation is often the performance bottleneck, as is the case for scraping `nixpkgs` during `pkgdb scrape`. Using the cache in this way means that installing a package built for one environment is nearly instantaneous for a different environment.

```
chonker:pkgdb zmitchell$ time flox install ripgrep
✅ 'ripgrep' installed to environment pkgdb at /Users/zmitchell/code/floxdev/flox/flox-pull-extra-builds/pkgdb

real	0m0.448s
user	0m0.276s
sys	0m0.082s
```

Similarly, if you pull an environment, remove it, and pull it again, the second pull is very quick because the store path for the previous build already exists on the system (from the first build) and the evaluation of all the constituent packages has already been cached.

```
chonker:pkgdb zmitchell$ flox pull ghudgins/default
✨ Pulled ghudgins/default from https://hub.flox.dev/

You can activate this environment with 'flox activate'

chonker:pkgdb zmitchell$ rm -rf .flox
chonker:pkgdb zmitchell$ time flox pull ghudgins/default
✨ Pulled ghudgins/default from https://hub.flox.dev/

You can activate this environment with 'flox activate'


real	0m0.933s
user	0m0.242s
sys	0m0.162s
```

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
Most operations will be substantially faster.

<!-- Many thanks! -->